Harmonize with exported Boolean type.
Some platforms, most notably CUDA and wasm, don't have as well-defined
ABI as more traditional platforms, and reinterpreting 'limb_t' or
'size_t' as 'bool' can be problematic. 'int' is more suitable for the
purpose. Just in case for reference, rationale behind this is to
leave compiler as little room for generating non-constant-time code
as possible.