ec_mult.h,multi_scalar.c: make array of pointer adaptive.
If you initialize a pointer in array of pointers to NULL, then previous
pointer is treated as pointer to a flat array of objects. This is done
to excuse caller from collecting pointers to individual objects from a
flat array. In other words, if you have a flat array of objects, then
you can pass it as ptr_array[2] = { array, NULL }, yet denote the actual
number of elements in the array.