aggregate.c: add blst_fp12_finalverify.
blst_fp12_finalverify makes it simpler to use blst in other pairing-based
schemes. Common pairing motif is

    C1 = e(A, B)
    C2 = e(C, D)
    return C1 == C2.

However, it's possible to optimize it by replacing e() with a call that
returns intermediate result and then performing the equivalent of the
comparison on these intermediate results. This latter step is what
blst_fp12_finalverify does. In other words the above procedure would
look as following:

    blst_miller_loop(C1, A, B);
    blst_miller_loop(C2, C, D);
    return blst_fp12_finalverify(C1, C2);