asm/ct_inverse_mod_256-armv8.pl: switch to k=31.
The choice of the k parameter is a matter of delicate balance between
how many bits are generated in the 'inner loop' and how fast
multiplications are. Cortex cores happen to be "in 256-bit balance"
in sense that performance is the same for both k=62 and k=31. This
makes k=31 preferred choice, because it would benefit non-Cortex cores.