bindings/go/blst_px.tgo: re-think multi-point/scalar interfaces.
In addition this commit also brings in unsafe.Pointer-s. This is because
of the "cgo argument has Go pointer to Go pointer" runtime error/panic.
It's possible to bypass it by setting the GODEBUG environment variable
to "cgocheck=0", but we can't rely on users to do so. It's argued that
it's safe to use unsafe.Pointer-s under the circumstances, because
C subroutines in question commit to not changing any of the pointers.