bindings/go/blst_px.tgo: handle pointer to pointer switcheroo differently.
Rationale is that _cgoCheckPointer suppression is as controversial as
uintptr(unsafe.Pointer), but yields more readable and go-vet-able code.
As a reminder, it's argued that under the circumstances it's safe to
perform the pointer to pointer switcheroo, because C subroutines in
question commit to not changing any of the pointers.