{no_asm,vect}.h: work around implicit boolean casts in clang>=14.
clang>=14 manages to identify that some limbs are assigned 0 or 1 values
and allows itself to use them as inputs to 'select' llvm bytecode, as if
the values are of actual C boolean type. This is problematic, because it
allows the final machine code generator to generate non-constant-time
code. And the latter is not speculative, the compiler in question was
observed to emit problematic code for no-asm versions of at least[!]
ptype##_gather_booth_w##SZ, fundamental to security of the secret-key
operations, and final subtraction in Montgomery multiplication...

Essential to note that x86_64 and arm64 are *not* actually affected, and
expectedly so, because critical subroutines are implemented in assembly.
Other than that it's as essential to recognize that the workaround
merely helps through the day, it's not a sustainable solution. Because
there is no guarantee that the compiler won't see through it tomorrow.
Solution would be a wider platform assembly coverage.