Change scratch register r8 -> rdi for LVI hardening in __inner_loop_62
Currently if LVI_HARDENING is turned on, r8(a_lo) will be used as a scratch register for __inner_loop_62's return.
When that happens, the a_lo value will be cobbled and on line 455, the cobbled a_lo value will be saved into out_ptr
which will be used in subsequent computations resulting in incorrect results if inp = {0,0,0}.

This commit changes the scratch register from r8 to rdi(cnt). rdi is used as a counter for __inner_loop_62 and is not
used in any subsequent code in ctq_inverse_mod_384.